+++ src/dup_utils/core/base/elements.py
@@ -1,17 +1,24 @@
+from __future__ import annotations
+
 import sys
 from collections import deque
 from collections.abc import Mapping, Set
 from numbers import Number
-from typing import Any, Optional
+from typing import Any, Optional, List, Dict
 
-from src.dup_utils.core.base.merge_split import split
+from src.dup_utils.core.base.split import split
 
 ZERO_DEPTH_BASES = (str, bytes, Number, range, bytearray)
 
 
-def only_one(check: list, value: list, default: bool = True) -> Optional[str]:
-    """Get only one element in check list that exists in match list
-    :usage:
+def only_one(
+        check: List[str],
+        value: List[str],
+        default: bool = True,
+) -> Optional[str]:
+    """Get only one element from check list that exists in match list.
+
+    .. usage:
         >>> only_one(['a', 'b'], ['a', 'b', 'c'])
         'a'
         >>> only_one(['a', 'b'], ['c', 'e', 'f'])
@@ -27,15 +34,14 @@ def only_one(check: list, value: list, default: bool = True) -> Optional[str]:
     )
 
 
-def hasdot(search: str, content: dict) -> bool:
-    """
-    :usage:
+def hasdot(search: str, content: Dict[Any, Any]) -> bool:
+    """Return True value if dot searching exists in content data.
+
+    .. usage:
         >>> hasdot('data.value', {'data': {'value': 2}})
         True
-
         >>> hasdot('data.value.key', {'data': {'value': 2}})
         False
-
         >>> hasdot('item.value.key', {'data': {'value': 2}})
         False
     """
@@ -50,23 +56,21 @@ def hasdot(search: str, content: dict) -> bool:
 
 def getdot(
     search: str,
-    content: dict,
+    content: Dict[Any, Any],
     *args,
     **kwargs,
 ) -> Any:
-    """
-    :usage:
+    """Return the value if dot searching exists in content data.
+
+    .. usage:
         >>> getdot('data.value', {'data': {'value': 1}})
         1
-
         >>> getdot('data', {'data': 'test'})
         'test'
-
         >>> getdot('data.value', {'data': 'test'})
         Traceback (most recent call last):
         ...
         ValueError: 'value' does not exists in test
-
         >>> getdot('data.value', {'data': {'key': 1}}, None)
 
         >>> getdot(
@@ -74,7 +78,12 @@ def getdot(
         ...     {'data': {'value': {'getter': 'success', 'put': 'fail'}}},
         ... )
         'success'
+        >>> getdot('foo.bar', {"foo": {"baz": 1}}, ignore=True)
 
+        >>> getdot('foo.bar', {"foo": {"baz": 1}}, 2, 3)
+        2
+        >>> getdot('foo.bar', {"foo": {"baz": 1}}, 2, 3, ignore=True)
+        2
     """
     _ignore: bool = kwargs.get("ignore", False)
     _search, _else = split(search, ".", maxsplit=1)
@@ -93,12 +102,13 @@ def getdot(
     raise ValueError(f"{_search} does not exists in {content}")
 
 
-def setdot(search: str, content: dict, value: Any, **kwargs) -> dict:
+def setdot(
+        search: str, content: dict, value: Any, **kwargs
+) -> Dict:
     """
-    :usage:
+    .. usage:
         >>> setdot('data.value', {'data': {'value': 1}}, 2)
         {'data': {'value': 2}}
-
         >>> setdot('data.value.key', {'data': {'value': 1}}, 2, ignore=True)
         {'data': {'value': 1}}
     """
@@ -122,21 +132,28 @@ def setdot(search: str, content: dict, value: Any, **kwargs) -> dict:
 def size(value: Any) -> int:
     """Recursively iterate to sum size of object & members.
 
-    Empty
-    Bytes  type        scaling notes
-    28     int         +4 bytes about every 30 powers of 2
-    37     bytes       +1 byte per additional byte
-    49     str         +1-4 per additional character (depending on max width)
-    48     tuple       +8 per additional item
-    64     list        +8 for each additional
-    224    set         5th increases to 736; 21nd, 2272; 85th, 8416; 341, 32992
-    240    dict        6th increases to 368; 22nd, 1184; 43rd, 2280; 86th, 4704; 171st, 9320
-    136    func def    does not include default args and other attrs
-    1056   class def   no slots
-    56     class inst  has a __dict__ attr, same scaling as dict above
-    888    class def   with slots
-    16     __slots__   seems to store in mutable tuple-like structure
-                       first slot grows to 48, and so on.
+        Empty
+        Bytes  type        scaling notes
+        28     int         +4 bytes about every 30 powers of 2
+        37     bytes       +1 byte per additional byte
+        49     str         +1-4 per additional character (depending on max width)
+        48     tuple       +8 per additional item
+        64     list        +8 for each additional
+        224    set         5th increases to 736; 21nd, 2272; 85th, 8416; 341, 32992
+        240    dict        6th increases to 368; 22nd, 1184; 43rd, 2280; 86th,
+                            4704; 171st, 9320
+        136    func def    does not include default args and other attrs
+        1056   class def   no slots
+        56     class inst  has a __dict__ attr, same scaling as dict above
+        888    class def   with slots
+        16     __slots__   seems to store in mutable tuple-like structure
+                            first slot grows to 48, and so on.
+
+    .. usage:
+        >>> size({'foo': 'bar'})
+        336
+        >>> size('foo')
+        52
     """
     _seen_ids = set()